+++ torch/data.py
@@ -1,0 +1,183 @@
+# Helper functions for data preprocessing.
+
+import os
+import pandas as pd
+import numpy as np
+from datetime import datetime
+from sklearn.preprocessing import MinMaxScaler
+
+
+def create_output_dir(root):
+    rslt_dir = root / 'created_models'
+    if not os.path.exists(rslt_dir):
+        os.mkdir(rslt_dir)
+    
+    date_str = datetime.datetime.today().strftime('%Y-%m-%d_%H%M')
+    out_dir = rslt_dir / date_str
+    if not os.path.exists(out_dir):
+        os.mkdir(out_dir)
+    print('directory', out_dir, 'has been created.\n')
+    
+    return out_dir
+
+
+def data_preproc(data_table, lin=True):
+    scale_exp = []
+    trgt_params = ('potential (V)', 'Ne (#/m^-3)', 'Ar+ (#/m^-3)', 'Nm (#/m^-3)', 'Te (eV)')
+
+    def get_param_exp(col_vals):
+        ''' get exponent of the parameter's mean value for scaling. '''
+        mean_exp = round(np.log10(col_vals.mean()), 0) - 1.0
+        # return 0 if mean_exp is less than zero to avoid blowing up small values
+        if mean_exp >=  0.0:
+            return mean_exp
+        else:
+            return 0.0
+
+
+    for col_n,(col_name,col_vals) in enumerate(data_table.iteritems(), start=1):
+        if col_name in trgt_params:
+            if lin:
+                # get exponent for scaling
+                exponent = get_param_exp(col_vals)
+                scale_exp.append(exponent)            
+                tmp_col = col_vals.values.reshape(-1,1)/(10**exponent)  # scale by dividing
+            else:
+                tmp_col = np.log10(col_vals.values.reshape(-1,1))
+        else:
+            tmp_col = col_vals.values.reshape(-1,1)
+        proced_table = tmp_col if col_n==1 else np.hstack([proced_table,tmp_col])
+    
+    proced_table = pd.DataFrame(proced_table, columns=data_table.columns)
+    proced_table = proced_table.replace([np.inf,-np.inf], np.nan)
+    proced_table = proced_table.dropna(how='any')
+    
+    return proced_table
+
+
+def scale_all(data_table, x_or_y, out_dir=None):
+    data_table = data_table.copy()
+    for n,column in enumerate(data_table.columns, start=1):
+        scaler = MinMaxScaler()
+        data_col = data_table[column].values.reshape(-1,1)
+        scaler.fit(data_col)
+        scaled_data = scaler.transform(data_col)
+        scaled_data_table = scaled_data if n==1 else np.hstack([scaled_data_table,scaled_data])
+        
+        if out_dir is not None:
+            pickle_file = out_dir / f'{x_or_y}scaler_{n:02d}.pkl'
+            with open(pickle_file, mode='wb') as pf:
+                pickle.dump(scaler, pf, protocol=4)
+    
+    scaled_data_table = pd.DataFrame(scaled_data_table, columns=data_table.columns)
+    
+    return scaled_data_table
+
+def read_aug_data(file):
+    """Read data file and return a DataFrame.
+
+    Args:
+        file (PosixPath): Path to .feather file.
+
+    Returns:
+        interp_df: DataFrame of interpolated data.
+    """
+    interp_df = pd.read_feather(file).drop(columns=['Ex (V/m)', 'Ey (V/m)'])
+    return interp_df
+
+
+##### data processing ######
+def get_augmentation_data(data_used, xy: bool, vp: bool):
+    # TODO: move to data module
+    """Get augmentation data for training.
+
+    Args:
+        xy (bool): Include xy grid augmented data
+        vp (bool): Inclde vp augmentation data
+    """
+    if xy:  # xy augmentation
+        xyfile = Path(root/'data'/'interpolation_datasets'/'rec-interpolation2.nc')
+        xydf = xr.open_dataset(xyfile).to_dataframe().reset_index().dropna()
+    else: xydf = None
+
+    if vp:  # vp augmentation
+        vpfolder = Path(root/'data'/'interpolation_feather'/'20221209')
+        # read all files and combine into a single df
+        vpdf = [read_aug_data(file) for file in vpfolder.glob('*.feather')]
+        vpdf = pd.concat(vpdf).rename(columns={'Vpp [V]' : 'V', 
+                                               'P [Pa]'  : 'P',
+                                               'X'       : 'x',
+                                               'Y'       : 'y'}, inplace=True)
+    else: vpdf = None
+
+    # make sure that the data follows the correct format before returning
+    assert list(data_used.columns) == ['V', 'P', 'x', 'y'] + label_names
+    return pd.concat([data_used, xydf, vpdf], ignore_index=True)
+    
+
+def get_data(xy=False, vp=False):
+    # TODO: move to data module
+    """Get dataset
+
+    Assumes the DataFrame has previously been saved as a .feather file. If not,
+    a new .feather file is created in the root/data folder.
+
+    Args:
+        xy (bool, optional): Include xy augmentation. Defaults to False.
+        vp (bool, optional): Include vp augmentation. Defaults to False.
+
+    Raises:
+        Exception: Raises an error if no data is available in avg_data.
+
+    Returns:
+        data_used: DataFrame of data to be used, with specified augmentation data.
+        data_excluded: DataFrame of excluded data.
+    """
+    avg_data_file = root/'data'/'avg_data.feather'
+
+    # check if feather file exists and load avg_data
+    if avg_data_file.is_file():
+        print('reading data feather file...')
+        avg_data = pd.read_feather(avg_data_file)
+    
+    else:
+        print('data feather file not found, building file...')
+        start_time = time.time()
+        avg_data = data.read_all_data(data_fldr_path, voltages, pressures).drop(columns=['Ex (V/m)', 'Ey (V/m)'])
+        elapsed_time = time.time() - start_time
+        print(f' done ({elapsed_time:0.1f} sec).\n')
+
+        if len(avg_data)==0:
+            raise Exception('No data available.')
+
+        avg_data.to_feather(avg_data_file)
+
+    # separate data to be excluded (to later check the model)
+    data_used     = avg_data[~((avg_data['Vpp [V]']==voltage_excluded) & (avg_data['P [Pa]']==pressure_excluded))].copy()
+    data_excluded = avg_data[  (avg_data['Vpp [V]']==voltage_excluded) & (avg_data['P [Pa]']==pressure_excluded) ].copy()
+
+    # rename columns
+    data_used.rename(columns={'Vpp [V]' : 'V',
+                              'P [Pa]'  : 'P',
+                              'X'       : 'x', 
+                              'Y'       : 'y'}, inplace=True)
+
+    if (xy or vp):
+        data_used = get_augmentation_data(data_used, xy, vp)
+
+    # create new column of x^2 and y^2
+    data_used['x**2'] = data_used['x']**2
+    data_used['y**2'] = data_used['y']**2
+    
+    return data_used, data_excluded
+
+##### misc #####
+def yn(str):
+    # TODO: move to data module
+    if str.lower() in ['y', 'yes', 'yea', 'ok', 'okay', 'k',  
+                       'sure', 'hai', 'aye', 'ayt', 'fosho']:
+        return True
+    elif str.lower() in ['n', 'no', 'nope', 'nah', 'hold this l']:
+        return False
+    else:
+        raise Exception(str + 'not recognized: use y - yes, n - no')
